updat docs for side sheet and navigation add back contextual to top app bar
