--- src-docs/public/top-app-bar-prominent.html
@@ -1,91 +1,0 @@
-<html>
-<head>
-  <meta http-equiv="Cache-Control" content="no-store" />
-  <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
-  <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Fira+Mono" />
-  <link href="/pax-components/theme.css" rel="stylesheet">
-  <link href="/pax-components/entry.css" rel="stylesheet">
-  <script src="/pax-components/entry.js"></script>
-  <link rel="stylesheet" href="main.css">
-</head>
-
-  <body>
-    <mdw-page>
-      <header>
-        <mdw-top-app-bar mdw-prominent mdw-shrink>
-          <section mdw-fixed>
-            <sub-section>
-              <mdw-icon>menu</mdw-icon>
-            </sub-section>
-          </section>
-          
-          <section mdw-flex>
-            <sub-section></sub-section>
-            <sub-section>
-              <span
-                class="mdw-title"
-                mdw-animation-property="transform: scale(#)"
-                mdw-animation-start="1.5"
-                mdw-animation-end="1"
-              >Title</span>
-            </sub-section>
-          </section>
-          
-          <section mdw-fixed>
-            <sub-section>
-              <mdw-icon>more_vert</mdw-icon>
-            </sub-section>
-          </section>
-        </mdw-top-app-bar>
-      </header>
-
-      <mdw-content>
-          <p>
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit,
-    sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
-    enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-    aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in
-    voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
-    occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
-    anim id est laborum.
-          </p>
-          <p>
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit,
-    sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
-    enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-    aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in
-    voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
-    occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
-    anim id est laborum.
-          </p>
-          <p>
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit,
-    sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
-    enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-    aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in
-    voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
-    occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
-    anim id est laborum.
-          </p>
-          <p>
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit,
-            sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
-            enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-            aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in
-            voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
-            occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
-            anim id est laborum.
-          </p>
-          <p>
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit,
-            sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
-            enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-            aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in
-            voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
-            occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
-            anim id est laborum.
-          </p>
-      </mdw-content>
-    </mdw-page>
-  </body>
-</html>
